back/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
psycopg2
psycopg2-binary